refactor(controller): simplify program scheduling metadata
