+++ crates/arroyo-controller/src/states/scheduling.rs
@@ -708,84 +708,3 @@ impl State for Scheduling {
         ))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn protobuf_scheduling_preserves_task_placement() {
-        let program = api::ArrowProgram {
-            nodes: [(0, 10, 2), (1, 20, 3), (2, 30, 1)]
-                .into_iter()
-                .map(|(node_index, node_id, parallelism)| api::ArrowNode {
-                    node_index,
-                    node_id,
-                    parallelism,
-                    ..Default::default()
-                })
-                .collect(),
-            ..Default::default()
-        };
-        let workers: Vec<_> = (0..2)
-            .map(|id| WorkerStatus {
-                id: WorkerId(id),
-                machine_id: MachineId(Arc::new(format!("machine-{id}"))),
-                rpc_address: format!("rpc-{id}"),
-                data_address: format!("data-{id}"),
-                slots: 2,
-                state: WorkerState::Connected,
-            })
-            .collect();
-
-        for (overrides, expected) in [
-            (
-                HashMap::new(),
-                vec![
-                    (10, 0, 0),
-                    (10, 1, 0),
-                    (20, 0, 0),
-                    (20, 1, 0),
-                    (20, 2, 1),
-                    (30, 0, 0),
-                ],
-            ),
-            (
-                HashMap::from([(10, 4), (30, 2)]),
-                vec![
-                    (10, 0, 0),
-                    (10, 1, 0),
-                    (10, 2, 1),
-                    (10, 3, 1),
-                    (20, 0, 0),
-                    (20, 1, 0),
-                    (20, 2, 1),
-                    (30, 0, 0),
-                    (30, 1, 0),
-                ],
-            ),
-        ] {
-            let mut scheduled = program.clone();
-            scheduled.update_parallelism(&overrides).unwrap();
-            let assignments = compute_assignments(workers.iter().collect(), &scheduled);
-            assert_eq!(assignments.len(), scheduled.task_count());
-            assert_eq!(
-                assignments
-                    .iter()
-                    .map(|a| (a.task_id, a.subtask_idx, a.worker_id))
-                    .collect::<Vec<_>>(),
-                expected
-            );
-            for assignment in assignments {
-                assert_eq!(
-                    assignment.worker_addr,
-                    format!("data-{}", assignment.worker_id)
-                );
-                assert_eq!(
-                    assignment.worker_rpc,
-                    format!("rpc-{}", assignment.worker_id)
-                );
-            }
-        }
-    }
-}
